Campus_Insight/requirements.txt
django
django-allauth
requests
jwt
python-docx
